Initial Code and Documentation
The cardinal dynamic nature of a discrete event simulation and its
instantaneous state changes
at occurance of specific events, make it difficult for the distribution
of the model output values
to attain a constant, steady behaviour; especially, at the very start of
the simulation run. The
presence of this prominent bias at the beginning of the simulation leads
to an imprecise estimation
of the system parameters.Removing the transient period by estimating the
appropriate truncation
point and looking for the start of the steady state, ignoring the
transient phase have been the
two promising solutions to this start-up bias, but however, most
approaches from the literature
have been subjective. A transient solution algorithm that can be
evaluated for its performance
has to be more comprehensive.
The intent of this master thesis is to find a method that can achieve
the best accuracy out of any
simulation run. There are many methods that have already been documented
and the many other
which might be suggested for the dealing with the initial transient.
However, there are no pre-fixed
parameters for quantizing the evaluation of these methods. One of the
main deliverable from this
work is a software framework, that provides an interface for comparing
various different truncation
methods with each other. The integrity of the evaluated algorithms is
the output obtained from
the developed interface, and is decided by running the heuristics
against the different benchmarks,
with random number seeds.
A variety of methods which included online methods and offline -
converted into online methods were implemented in the current framework
and tested with different model data that was
obtained from TimeNET. The final artifacts of this work include an
executable framework, the
comparison between different methods deduced from the results given by
the framework.